moved BaseRetriever into its own class to prevent circular dependency
