+++ langchain/src/chains/retrieval_qa.ts
@@ -1,7 +1,8 @@
 import { BaseChain, ChainInputs } from "./base.js";
 import { BaseLanguageModel } from "../base_language/index.js";
 import { SerializedVectorDBQAChain } from "./serde.js";
-import { ChainValues, BaseRetriever } from "../schema/index.js";
+import { BaseRetriever } from "../retrievers/base.js";
+import { ChainValues } from "../schema/index.js";
 import {
   StuffQAChainParams,
   loadQAStuffChain,
