+++ langchain/src/retrievers/contextual_compression.ts
@@ -1,6 +1,6 @@
 import { BaseDocumentCompressor } from "./document_compressors/index.js";
 import { Document } from "../document.js";
-import { BaseRetriever } from "../schema/index.js";
+import { BaseRetriever } from "./base.js";
 
 export interface ContextualCompressionRetrieverArgs {
   baseCompressor: BaseDocumentCompressor;
