+++ langchain/src/retrievers/databerry.ts
@@ -1,4 +1,4 @@
-import { BaseRetriever } from "../schema/index.js";
+import { BaseRetriever } from "./base.js";
 import { Document } from "../document.js";
 import { AsyncCaller, AsyncCallerParams } from "../util/async_caller.js";
